requirements.txt
asgiref==3.11.0
Django==5.0.14
mssql-django==1.6
pyodbc==5.3.0
sqlparse==0.5.4
reportlab==4.0.7
arabic-reshaper==3.0.0
python-bidi==0.4.2
ollama==0.4.4
waitress==3.0.2
